Started work on a new world which is a forge world - lakes of liquid methane and great veins of molten metal
Issues are persisting with the temperature, methane won't stay cold and the iron and copper tubes aren't spawing correctly - maybe move them from spawning akin to the megma vents to how spawning as a regular biome? Think the temperature thing is a bug, though. Also the metal tubes need to have a proper density being set, as it is it all just sinks into a puddle in the bottom. -Might be the override for the massMultiplier? Has to be tested, but not when I'm tired.